dnsdist: Split the DynBlocks regression tests
On my 8 physical CPU cores machine this brings the total time for
all regression tests to 107.60s (01:47).
